refactor(data-studio): clean up from migration to @leav/ui + new eyes on code
