+++ apps/data-studio/src/components/Navigation/Navigation.tsx
@@ -10,7 +10,7 @@ import {useTranslation} from 'react-i18next';
 import {setInfoBase} from 'reduxStore/infos';
 import {useAppDispatch, useAppSelector} from 'reduxStore/store';
 import {GET_TREES, GET_TREESVariables} from '_gqlTypes/GET_TREES';
-import {useActiveTree} from '../../hooks/ActiveTreeHook/ActiveTreeHook';
+import {useActiveTree} from 'hooks/useActiveTree/useActiveTree';
 import {isTreeInApp, localizedTranslation} from '../../utils';
 import {IBaseInfo, InfoType, WorkspacePanels} from '../../_types/types';
 import NavigationView from './NavigationView';
