+++ apps/data-studio/src/components/Navigation/NavigationView/Column/HeaderColumnNavigation/HeaderColumnNavigationActions/DefaultActions/DefaultActions.tsx
@@ -17,7 +17,7 @@ import {Button, Dropdown, message} from 'antd';
 import {ItemType} from 'antd/lib/menu/hooks/useItems';
 import {IconEllipsisVertical} from 'assets/icons/IconEllipsisVertical';
 import {removeTreeElementMutation} from 'graphQL/mutations/trees/removeTreeElementMutation';
-import {useActiveTree} from 'hooks/ActiveTreeHook/ActiveTreeHook';
+import {useActiveTree} from 'hooks/useActiveTree/useActiveTree';
 import useRefreshTreeContent from 'hooks/useRefreshTreeContent';
 import {useState} from 'react';
 import {useTranslation} from 'react-i18next';
