+++ apps/data-studio/src/components/Navigation/NavigationView/Column/HeaderColumnNavigation/HeaderColumnNavigationActions/HeaderColumnNavigationActions.tsx
@@ -2,7 +2,7 @@
 // This file is released under LGPL V3
 // License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
 import {Button} from 'antd';
-import {useActiveTree} from 'hooks/ActiveTreeHook/ActiveTreeHook';
+import {useActiveTree} from 'hooks/useActiveTree/useActiveTree';
 import {useTreeLibraryAllowedAsChild} from 'hooks/useTreeLibraryAllowedAsChild';
 import {useTranslation} from 'react-i18next';
 import {addInfo} from 'reduxStore/infos';
