+++ apps/data-studio/src/components/Sidebar/Sidebar.tsx
@@ -11,8 +11,8 @@ import AppIcon from 'components/shared/AppIcon';
 import TreeIcon from 'components/shared/TreeIcon';
 import {saveUserData} from 'graphQL/mutations/userData/saveUserData';
 import {getUserDataQuery} from 'graphQL/queries/userData/getUserData';
-import {useActiveLibrary} from 'hooks/ActiveLibHook/ActiveLibHook';
-import {useActiveTree} from 'hooks/ActiveTreeHook/ActiveTreeHook';
+import {useActiveLibrary} from 'hooks/useActiveLibrary/useActiveLibrary';
+import {useActiveTree} from 'hooks/useActiveTree/useActiveTree';
 import {useApplicationLibraries} from 'hooks/useApplicationLibraries';
 import {useApplicationTrees} from 'hooks/useApplicationTrees';
 import {useTranslation} from 'react-i18next';
