+++ apps/data-studio/src/constants/constants.ts
@@ -1,6 +1,6 @@
 // Copyright LEAV Solutions 2017
 // This file is released under LGPL V3
 // License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
-export const defaultInfosTime = 5000;
+export const defaultInfosTime = 5_000;
 
 export const treeNavigationPageSize = 20;
