--- apps/data-studio/src/context/selectionModeContext.ts
@@ -1,8 +1,0 @@
-// Copyright LEAV Solutions 2017
-// This file is released under LGPL V3
-// License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
-import {createContext} from 'react';
-
-const SelectionModeContext = createContext<boolean>(false);
-
-export default SelectionModeContext;
