+++ apps/data-studio/src/hooks/useActiveLibrary/useActiveLibrary.test.tsx
@@ -6,9 +6,9 @@ import {LibraryBehavior} from '_gqlTypes/globalTypes';
 import {render} from '_tests/testUtils';
 import {mockLibraryPermissions} from '__mocks__/common/library';
 import {IActiveLibrary} from '../../graphQL/queries/cache/activeLibrary/getActiveLibraryQuery';
-import {initialActiveLibrary, useActiveLibrary} from './ActiveLibHook';
+import {initialActiveLibrary, useActiveLibrary} from './useActiveLibrary';
 
-describe('ActiveLibHook', () => {
+describe('useActiveLibrary', () => {
     const mockActiveLibrary: IActiveLibrary = {
         id: 'test',
         name: 'test',
