+++ apps/data-studio/src/hooks/useGetLibraryDetailExtendedQuery/index.ts
@@ -1,6 +1,6 @@
 // Copyright LEAV Solutions 2017
 // This file is released under LGPL V3
 // License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
-import SelectionModeContext from './selectionModeContext';
+import {useGetLibraryDetailExtendedQuery} from './useGetLibraryDetailExtendedQuery';
 
-export {SelectionModeContext};
+export default useGetLibraryDetailExtendedQuery;
