--- apps/data-studio/src/hooks/useNotification/index.tsx
@@ -1,4 +1,0 @@
-// Copyright LEAV Solutions 2017
-// This file is released under LGPL V3
-// License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
-export {default} from './useNotification';
