+++ apps/data-studio/src/index.css
@@ -1,6 +1,3 @@
-/* @import '~antd/dist/antd.less'; */
-/* @import '~antd/dist/antd.less'; */
-
 body {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
@@ -20,9 +17,8 @@ code {
   max-height: 100vh;
   overflow: hidden;
 
-  background: none no-repeat;
   background-size: contain;
-  background-position: bottom;
+  background: none no-repeat bottom;
 }
 
 .ant-table-filter-dropdown-btns {
