+++ libs/ui/src/_utils/typeguards.ts
@@ -1,3 +1,6 @@
+// Copyright LEAV Solutions 2017
+// This file is released under LGPL V3
+// License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
 import {WithTypename} from '@leav/utils';
 
 export const hasTypename = <T>(value: any): value is WithTypename<T> => '__typename' in value;
