+++ libs/ui/src/components/RecordEdition/EditRecordContent/reducers/standardFieldReducer/useStandardFieldReducer.ts
@@ -1,3 +1,6 @@
+// Copyright LEAV Solutions 2017
+// This file is released under LGPL V3
+// License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
 import {useContext} from 'react';
 import {StandardFieldReducerContext} from './standardFieldReducerContext';
